website/pyproject.toml
[tool.poetry]
name = "website"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.2.8"
django-allauth = "^0.45.0"
psycopg2 = "^2.9.1"
python-dotenv = "^0.19.1"
requests = "^2.26.0"
Scrapy = "^2.5.1"
crochet = "^2.0.0"
PyMuPDF = "^1.19.3"
djangorestframework = "^3.13.0"
docx2txt = "^0.8"
PyJWT = "^2.3.0"
django-cors-headers = "^3.10.1"
djangorestframework-simplejwt = "^5.0.0"
google-api-python-client = "^2.33.0"
oauth2 = "^1.9.0"
google-auth = "^2.3.3"

[tool.poetry.dev-dependencies]
autopep8 = "^1.6.0"

[tool.poetry.scripts]
server = "website.scripts:django_run_server"
django = "website.scripts:django_run_manage"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
